feat: CarouselModalPopUPComponent 새로운 형식
