+++ iOS/Catalog/Catalog/ModelTypeSelection/Model/ModeTypeModel.swift
@@ -33,17 +33,7 @@ enum ModelTypeModel: Equatable {
 
 }
 
-struct ModelTypeDetailState: Equatable, Identifiable {
-  
-  var id: Int
-  var title: String
-  var description: String?
-  var choiceRatio: CLNumber?
-  var imageURL: URL?
-  var price: CLNumber
-  var hmgData: HMGModelTypeState?
-  
-}
+
 
 struct ModelTypeOptionState: Equatable, Hashable {
